delete test_frompresent.py
